feat: SSR / CSR 분리-#132
